+++ packages/api/Common/Counters/CounterKeys.cs
@@ -1,7 +1,0 @@
-namespace Api.Common.Counters;
-
-public static class CounterKeys
-{
-    public const string Customer = "customer";
-    public const string Contract = "contract";
-}