+++ packages/api/Common/Database/AppDbContext.cs
@@ -1,4 +1,3 @@
-using Api.Common.Counters;
 using Api.Modules.Contracts;
 using Api.Modules.Customers;
 using Api.Modules.LeaveTypes;
@@ -20,7 +19,6 @@ public class AppDbContext : DbContext
     public DbSet<ContractTask> ContractTasks => Set<ContractTask>();
     public DbSet<LeaveType> LeaveTypes => Set<LeaveType>();
     public DbSet<UserLeaveAllowance> UserLeaveAllowances => Set<UserLeaveAllowance>();
-    public DbSet<Counter> Counters => Set<Counter>();
 
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
@@ -28,7 +26,6 @@ public class AppDbContext : DbContext
         modelBuilder.ApplyConfiguration(new UserConfiguration());
         modelBuilder.ApplyConfiguration(new ContractConfiguration());
         modelBuilder.ApplyConfiguration(new ContractTaskConfiguration());
-        modelBuilder.ApplyConfiguration(new CounterConfiguration());
         modelBuilder.ApplyConfiguration(new LeaveTypeConfiguration());
         modelBuilder.ApplyConfiguration(new UserLeaveAllowanceConfiguration());
     }